requirements.txt
psycopg2-binary==2.8.4
pandas==1.0.3
scikit-learn==0.24.2
xgboost==1.4.2
flask
flask-restful